+++ sdk/src/jupiter/jupiterClient.ts
@@ -41,8 +41,179 @@ export interface Route {
 	swapMode: SwapMode;
 }
 
+/**
+ * 
+ * @export
+ * @interface RoutePlanStep
+ */
+export interface RoutePlanStep {
+  /**
+   * 
+   * @type {SwapInfo}
+   * @memberof RoutePlanStep
+   */
+  swapInfo: SwapInfo;
+  /**
+   * 
+   * @type {number}
+   * @memberof RoutePlanStep
+   */
+  percent: number;
+}
+
+export interface SwapInfo {
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  ammKey: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  label?: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  inputMint: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  outputMint: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  inAmount: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  outAmount: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  feeAmount: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof SwapInfo
+   */
+  feeMint: string;
+}
+
+/**
+ * 
+ * @export
+ * @interface PlatformFee
+ */
+export interface PlatformFee {
+  /**
+   * 
+   * @type {string}
+   * @memberof PlatformFee
+   */
+  amount?: string;
+  /**
+   * 
+   * @type {number}
+   * @memberof PlatformFee
+   */
+  feeBps?: number;
+}
+
+/**
+ * 
+ * @export
+ * @interface QuoteResponse
+ */
+export interface QuoteResponse {
+  /**
+   * 
+   * @type {string}
+   * @memberof QuoteResponse
+   */
+  inputMint: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof QuoteResponse
+   */
+  inAmount: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof QuoteResponse
+   */
+  outputMint: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof QuoteResponse
+   */
+  outAmount: string;
+  /**
+   * 
+   * @type {string}
+   * @memberof QuoteResponse
+   */
+  otherAmountThreshold: string;
+  /**
+   * 
+   * @type {SwapMode}
+   * @memberof QuoteResponse
+   */
+  swapMode: SwapMode;
+  /**
+   * 
+   * @type {number}
+   * @memberof QuoteResponse
+   */
+  slippageBps: number;
+  /**
+   * 
+   * @type {PlatformFee}
+   * @memberof QuoteResponse
+   */
+  platformFee?: PlatformFee;
+  /**
+   * 
+   * @type {string}
+   * @memberof QuoteResponse
+   */
+  priceImpactPct: string;
+  /**
+   * 
+   * @type {Array<RoutePlanStep>}
+   * @memberof QuoteResponse
+   */
+  routePlan: Array<RoutePlanStep>;
+  /**
+   * 
+   * @type {number}
+   * @memberof QuoteResponse
+   */
+  contextSlot?: number;
+  /**
+   * 
+   * @type {number}
+   * @memberof QuoteResponse
+   */
+  timeTaken?: number;
+}
+
 export class JupiterClient {
-	url = 'https://quote-api.jup.ag/v4';
+	url = 'https://quote-api.jup.ag';
 	connection: Connection;
 	lookupTableCahce = new Map<string, AddressLookupTableAccount>();
 
@@ -51,6 +222,7 @@ export class JupiterClient {
 	}
 
 	/**
+	 * ** @deprecated - use getQuote
 	 * Get routes for a swap
 	 * @param inputMint the mint of the input token
 	 * @param outputMint the mint of the output token
@@ -84,13 +256,86 @@ export class JupiterClient {
 		}).toString();
 
 		const { data: routes } = await (
-			await fetch(`https://quote-api.jup.ag/v4/quote?${params}`)
+			await fetch(`${this.url}/v4/quote?${params}`)
 		).json();
 
 		return routes;
 	}
 
 	/**
+	 * Get routes for a swap
+	 * @param inputMint the mint of the input token
+	 * @param outputMint the mint of the output token
+	 * @param amount the amount of the input token
+	 * @param slippageBps the slippage tolerance in basis points
+	 * @param swapMode the swap mode (ExactIn or ExactOut)
+	 * @param onlyDirectRoutes whether to only return direct routes
+	 */
+	public async getQuote({
+		inputMint,
+		outputMint,
+		amount,
+		slippageBps = 50,
+		swapMode = 'ExactIn',
+		onlyDirectRoutes = false,
+	}: {
+		inputMint: PublicKey;
+		outputMint: PublicKey;
+		amount: BN;
+		slippageBps?: number;
+		swapMode?: SwapMode;
+		onlyDirectRoutes?: boolean;
+	}): Promise<QuoteResponse> {
+		const params = new URLSearchParams({
+			inputMint: inputMint.toString(),
+			outputMint: outputMint.toString(),
+			amount: amount.toString(),
+			slippageBps: slippageBps.toString(),
+			swapMode,
+			onlyDirectRoutes: onlyDirectRoutes.toString(),
+		}).toString();
+		const quote = await (
+			await fetch(`${this.url}/v6/quote?${params}`)
+		).json();
+		return quote;
+	}
+
+	/**
+	 * Get a swap transaction for quote
+	 * @param quoteResponse quote to perform swap
+	 * @param userPublicKey the signer's wallet public key
+	 * @param slippageBps the slippage tolerance in basis points
+	 */
+	public async getSwap({
+		quote,
+		userPublicKey,
+		slippageBps = 50,
+	}: {
+		quote: QuoteResponse;
+		userPublicKey: PublicKey;
+		slippageBps?: number;
+	}): Promise<VersionedTransaction> {
+		const resp = await (
+			await fetch(`${this.url}/v6/swap`, {
+				method: 'POST',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				body: JSON.stringify({
+          quoteResponse: quote,
+					userPublicKey,
+					slippageBps,
+				}),
+			})
+		).json();
+		const { swapTransaction } = resp;
+
+		const swapTransactionBuf = Buffer.from(swapTransaction, 'base64');
+		return VersionedTransaction.deserialize(swapTransactionBuf);
+	}
+
+	/**
+	 * ** @deprecated - use getSwap
 	 * Get a swap transaction for a route
 	 * @param route the route to perform swap
 	 * @param userPublicKey the signer's wallet public key
@@ -106,7 +351,7 @@ export class JupiterClient {
 		slippageBps?: number;
 	}): Promise<VersionedTransaction> {
 		const resp = await (
-			await fetch(`${this.url}/swap`, {
+			await fetch(`${this.url}/v4/swap`, {
 				method: 'POST',
 				headers: {
 					'Content-Type': 'application/json',
